feat(rendering): Add jpg as an output render option
* Add ability to render output as jpeg
* Modify parameter integer types to avoid lossy conversions
* Add static casts where lossy conversions do need to occur
* Move libatomic to be a private link dependency

Closes #4.

+++ include/color.h
@@ -8,6 +8,6 @@
 using Color = Eigen::Vector3d;
 
 auto writeColor(Color rPixelColor) -> std::string;
-auto writeColor(Color rPixelColor, uint32_t rSamplesPerPixel) -> std::string;
+auto writeColor(Color rPixelColor, size_t rSamplesPerPixel) -> std::string;
 auto getRGB(Color rPixelColor) -> Color;
-auto getRGB(Color rPixelColor, uint32_t rSamplesPerPixel) -> Color;
+auto getRGB(Color rPixelColor, size_t rSamplesPerPixel) -> Color;
